requirements.txt
torch==2.10.0+cpu
torchvision==0.25.0+cpu
streamlit==1.52.1
pillow==12.0.0
matplotlib==3.10.7
